+++ courses/serializers/v1/programs.py
@@ -6,7 +6,7 @@ from rest_framework.exceptions import ValidationError
 
 from cms.serializers import ProgramPageSerializer
 from courses import models
-from courses.serializers import get_thumbnail_url, BaseProgramRequirementTreeSerializer
+from courses.serializers.base import get_thumbnail_url, BaseProgramRequirementTreeSerializer
 from courses.serializers.v1.departments import DepartmentSerializer
 from courses.serializers.v1.courses import CourseWithCourseRunsSerializer, CourseRunEnrollmentSerializer
 from courses.serializers.v1.base import CourseRunCertificateSerializer, CourseRunGradeSerializer
