--- courses/serializers/v1/serializers_test.py
@@ -1,10 +1,0 @@
-"""
-Tests for course serializers
-"""
-# pylint: disable=unused-argument, redefined-outer-name
-
-import pytest
-
-pytestmark = [pytest.mark.django_db]
-
-
