--- courses/serializers/v2/__init__.py
@@ -1,95 +1,0 @@
-"""
-Course model serializers - v2
-"""
-import logging
-
-from rest_framework import serializers
-
-from courses.models import Program, ProgramRequirementNodeType
-from courses.serializers import get_thumbnail_url, BaseProgramRequirementTreeSerializer
-from courses.serializers.v1 import ProgramPageSerializer
-from main.serializers import StrictFieldsSerializer
-
-logger = logging.getLogger(__name__)
-
-
-class ProgramSerializer(serializers.ModelSerializer):
-    """Program Model Serializer v2"""
-
-    courses = serializers.SerializerMethodField()
-    requirements = serializers.SerializerMethodField()
-    req_tree = serializers.SerializerMethodField()
-    page = serializers.SerializerMethodField()
-    departments = serializers.PrimaryKeyRelatedField(many=True, read_only=True)
-
-    def get_courses(self, instance):
-        return [course[0].id for course in instance.courses if course[0].live]
-
-    def get_requirements(self, instance):
-        return {
-            "required": [course.id for course in instance.required_courses],
-            "electives": [course.id for course in instance.elective_courses],
-        }
-
-    def get_req_tree(self, instance):
-        req_root = instance.get_requirements_root()
-
-        if req_root is None:
-            return []
-
-        return ProgramRequirementTreeSerializer(instance=req_root).data
-
-    def get_page(self, instance):
-        if hasattr(instance, "page"):
-            return ProgramPageSerializer(instance.page).data
-        else:
-            return {"feature_image_src": get_thumbnail_url(None)}
-
-    class Meta:
-        model = Program
-        fields = [
-            "title",
-            "readable_id",
-            "id",
-            "courses",
-            "requirements",
-            "req_tree",
-            "page",
-            "program_type",
-            "departments",
-            "live",
-        ]
-
-
-class ProgramRequirementDataSerializer(StrictFieldsSerializer):
-    """Serializer for ProgramRequirement data"""
-
-    node_type = serializers.ChoiceField(
-        choices=(
-            ProgramRequirementNodeType.OPERATOR,
-            ProgramRequirementNodeType.COURSE,
-        )
-    )
-    course = serializers.CharField(source="course_id", allow_null=True, default=None)
-    program = serializers.CharField(source="program_id", required=False)
-    title = serializers.CharField(allow_null=True, default=None)
-    operator = serializers.CharField(allow_null=True, default=None)
-    operator_value = serializers.CharField(allow_null=True, default=None)
-    elective_flag = serializers.BooleanField(allow_null=True, default=False)
-
-
-class ProgramRequirementSerializer(StrictFieldsSerializer):
-    """Serializer for a ProgramRequirement"""
-
-    id = serializers.IntegerField(required=False, allow_null=True, default=None)
-    data = ProgramRequirementDataSerializer()
-
-    def get_fields(self):
-        """Override because 'children' is a recursive structure"""
-        fields = super().get_fields()
-        fields["children"] = ProgramRequirementSerializer(many=True, default=[])
-        return fields
-
-
-class ProgramRequirementTreeSerializer(BaseProgramRequirementTreeSerializer):
-    child = ProgramRequirementSerializer()
